Partially implemented Swipe for next
Partially got the swipe for next feature done, theres a glitch where i have to swipe twice for nexr, i know the fix but not how to implement it yet, and there's animations left. after that there's the pre-load thing but i think i might start another project before that if i get a good idea.